similarity/requirements.txt
ImageHash==4.2.1
Pillow==9.2.0
